added glm finished inputinfolayer

+++ scaffold/include/Manifest.hpp
@@ -11,9 +11,9 @@ namespace Scaffold
         uint32_t initialWidth = 1280;
         uint32_t initialHeight = 720;
 
-        uint8_t swapInterval = 1;
+        uint32_t swapInterval = 1;
 
-        float eventWaitTimeout = -1;
+        float eventWaitTimeout = -1.0f;
 
         bool dockspaceOverViewport = false;
         bool useProfilerLayer = false;
